pond4/containers/back/requirements.txt
fastapi
uvicorn
pycaret
pandas
numpy
pydantic
unidecode
requests
streamlit
psycopg2-binary
PyJWT
